# Conflicts:
#	src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/provider/rbac/service/RbacPermissionResourceMemberService.kt
#	src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/service/AuthDeptServiceImpl.kt
#	src/backend/ci/core/openapi/biz-openapi/src/main/kotlin/com/tencent/devops/openapi/resources/apigw/v4/ApigwAtomResourceV4Impl.kt
#	src/backend/ci/core/openapi/biz-openapi/src/main/kotlin/com/tencent/devops/openapi/resources/apigw/v4/ApigwProjectResourceV4Impl.kt
#	src/backend/ci/core/project/api-project/src/main/kotlin/com/tencent/devops/project/api/service/ServiceProjectResource.kt
#	src/backend/ci/core/project/biz-project/src/main/kotlin/com/tencent/devops/project/resources/ServiceProjectResourceImpl.kt
#	src/backend/ci/core/project/biz-project/src/main/kotlin/com/tencent/devops/project/service/ProjectService.kt
#	src/backend/ci/core/project/biz-project/src/main/kotlin/com/tencent/devops/project/service/impl/AbsProjectServiceImpl.kt
#	src/backend/ci/core/store/api-store/src/main/kotlin/com/tencent/devops/store/api/atom/ServiceAtomResource.kt
#	src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/atom/resources/ServiceAtomResourceImpl.kt
#	src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/atom/service/impl/OpAtomServiceImpl.kt
#	src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/image/dao/ImageDao.kt
#	src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/image/resources/ServiceStoreImageResourceImpl.kt
#	src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/image/service/ImageService.kt

+++ src/backend/ci/core/stream/biz-stream/src/main/kotlin/com/tencent/devops/stream/trigger/parsers/PipelineDelete.kt
@@ -119,8 +119,9 @@ class PipelineDelete @Autowired constructor(
                 pipelineId = pipelineId,
                 branch = action.data.eventCommon.branch
             )
-
-            repoTriggerEventService.deleteRepoTriggerEvent(pipelineId)
+            if (action.data.eventCommon.branch == action.data.context.defaultBranch) {
+                repoTriggerEventService.deleteRepoTriggerEvent(pipelineId)
+            }
 
             val isFileEmpty =
                 if (null != filePath && action.data.eventCommon.branch != action.data.context.defaultBranch) {
